refactor(java): :recycle: test refactoring done to reduce test runtime
