+++ core-java/src/test/java/com/github/wasiqb/boyka/testng/web/saucedemo/pages/LoginPage.java
@@ -14,7 +14,7 @@
  * copies or substantial portions of the Software.
  */
 
-package com.github.wasiqb.boyka.testng.web.pages.saucedemo;
+package com.github.wasiqb.boyka.testng.web.saucedemo.pages;
 
 import static com.github.wasiqb.boyka.builders.Locator.buildLocator;
 import static org.openqa.selenium.By.id;
@@ -23,26 +23,35 @@ import com.github.wasiqb.boyka.builders.Locator;
 import lombok.Getter;
 
 /**
- * Cart page.
+ * Login page object for Sauce demo application.
  *
  * @author Wasiq Bhamla
- * @since 13-Mar-2022
+ * @since 24-Feb-2022
  */
 @Getter
-public class CartPage {
+public class LoginPage {
     /**
-     * Gets Cart page instance.
+     * Login page object.
      *
-     * @return Cart page instance
+     * @return {@link LoginPage}
      */
-    public static CartPage cartPage () {
-        return new CartPage ();
+    public static LoginPage loginPage () {
+        return new LoginPage ();
     }
 
-    private final Locator checkout = buildLocator ().web (id ("checkout"))
+    private final Locator loginBox    = buildLocator ().web (id ("login_button_container"))
+        .build ();
+    private final Locator loginButton = buildLocator ().web (id ("login-button"))
+        .parent (this.loginBox)
+        .build ();
+    private final Locator password    = buildLocator ().web (id ("password"))
+        .parent (this.loginBox)
+        .build ();
+    private final Locator username    = buildLocator ().web (id ("user-name"))
+        .parent (this.loginBox)
         .build ();
 
-    private CartPage () {
-        // Avoid explicit class initialisation
+    private LoginPage () {
+        // Avoid explicit class initialisation.
     }
 }
