+++ core-java/test-suites/testng-web-grid.xml
@@ -16,30 +16,13 @@
   -->
 
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Boyka Web Grid Suite" verbose="2">
+<suite name="Boyka Web Grid Suite" verbose="2" parallel="classes">
     <test name="Test Boyka web on Selenium Grid">
         <parameter name="appType" value="WEB"/>
         <parameter name="driverKey" value="test_selenium_grid"/>
-        <classes>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWeb">
-                <methods>
-                    <include name="testLogin"/>
-                    <include name="testAddToCart"/>
-                    <include name="testProductDetailsPage"/>
-                    <include name="testProductCartPage"/>
-                    <include name="testCheckoutStep1"/>
-                    <include name="testCheckoutStep2"/>
-                    <include name="testSignOut"/>
-                </methods>
-            </class>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWebActions">
-                <methods>
-                    <include name="testAcceptAlert"/>
-                    <include name="testAcceptConfirmAlert"/>
-                    <include name="testDismissConfirmAlert"/>
-                    <include name="testDismissPromptAlert"/>
-                </methods>
-            </class>
-        </classes>
+        <packages>
+            <package name="com.github.wasiqb.boyka.testng.web.saucedemo"/>
+            <package name="com.github.wasiqb.boyka.testng.web.theinternet"/>
+        </packages>
     </test>
 </suite>
