+++ core-java/test-suites/testng-web-local.xml
@@ -16,105 +16,37 @@
   -->
 
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Boyka Web Local Suite" verbose="2">
+<suite name="Boyka Web Local Suite" verbose="2" parallel="classes" thread-count="10">
     <test name="Test Boyka web on Local Chrome">
         <parameter name="appType" value="WEB"/>
         <parameter name="driverKey" value="test_local_chrome"/>
-        <classes>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWeb">
-                <methods>
-                    <include name="testLogin"/>
-                    <include name="testAddToCart"/>
-                    <include name="testProductDetailsPage"/>
-                    <include name="testProductCartPage"/>
-                    <include name="testCheckoutStep1"/>
-                    <include name="testCheckoutStep2"/>
-                    <include name="testSignOut"/>
-                </methods>
-            </class>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWebActions">
-                <methods>
-                    <include name="testAcceptAlert"/>
-                    <include name="testAcceptConfirmAlert"/>
-                    <include name="testDismissConfirmAlert"/>
-                    <include name="testDismissPromptAlert"/>
-                </methods>
-            </class>
-        </classes>
+        <packages>
+            <package name="com.github.wasiqb.boyka.testng.web.saucedemo"/>
+            <package name="com.github.wasiqb.boyka.testng.web.theinternet"/>
+        </packages>
     </test>
     <test name="Test Boyka web on Local Firefox">
         <parameter name="appType" value="WEB"/>
         <parameter name="driverKey" value="test_local_firefox"/>
-        <classes>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWeb">
-                <methods>
-                    <include name="testLogin"/>
-                    <include name="testAddToCart"/>
-                    <include name="testProductDetailsPage"/>
-                    <include name="testProductCartPage"/>
-                    <include name="testCheckoutStep1"/>
-                    <include name="testCheckoutStep2"/>
-                    <include name="testSignOut"/>
-                </methods>
-            </class>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWebActions">
-                <methods>
-                    <include name="testAcceptAlert"/>
-                    <include name="testAcceptConfirmAlert"/>
-                    <include name="testDismissConfirmAlert"/>
-                    <include name="testDismissPromptAlert"/>
-                </methods>
-            </class>
-        </classes>
+        <packages>
+            <package name="com.github.wasiqb.boyka.testng.web.saucedemo"/>
+            <package name="com.github.wasiqb.boyka.testng.web.theinternet"/>
+        </packages>
     </test>
     <test name="Test Boyka web on Local Edge">
         <parameter name="appType" value="WEB"/>
         <parameter name="driverKey" value="test_local_edge"/>
-        <classes>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWeb">
-                <methods>
-                    <include name="testLogin"/>
-                    <include name="testAddToCart"/>
-                    <include name="testProductDetailsPage"/>
-                    <include name="testProductCartPage"/>
-                    <include name="testCheckoutStep1"/>
-                    <include name="testCheckoutStep2"/>
-                    <include name="testSignOut"/>
-                </methods>
-            </class>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWebActions">
-                <methods>
-                    <include name="testAcceptAlert"/>
-                    <include name="testAcceptConfirmAlert"/>
-                    <include name="testDismissConfirmAlert"/>
-                    <include name="testDismissPromptAlert"/>
-                </methods>
-            </class>
-        </classes>
+        <packages>
+            <package name="com.github.wasiqb.boyka.testng.web.saucedemo"/>
+            <package name="com.github.wasiqb.boyka.testng.web.theinternet"/>
+        </packages>
     </test>
     <test name="Test Boyka web on Local Safari" enabled="false">
         <parameter name="appType" value="WEB"/>
         <parameter name="driverKey" value="test_local_safari"/>
-        <classes>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWeb">
-                <methods>
-                    <include name="testLogin"/>
-                    <include name="testAddToCart"/>
-                    <include name="testProductDetailsPage"/>
-                    <include name="testProductCartPage"/>
-                    <include name="testCheckoutStep1"/>
-                    <include name="testCheckoutStep2"/>
-                    <include name="testSignOut"/>
-                </methods>
-            </class>
-            <class name="com.github.wasiqb.boyka.testng.web.TestWebActions">
-                <methods>
-                    <include name="testAcceptAlert"/>
-                    <include name="testAcceptConfirmAlert"/>
-                    <include name="testDismissConfirmAlert"/>
-                    <include name="testDismissPromptAlert"/>
-                </methods>
-            </class>
-        </classes>
+        <packages>
+            <package name="com.github.wasiqb.boyka.testng.web.saucedemo"/>
+            <package name="com.github.wasiqb.boyka.testng.web.theinternet"/>
+        </packages>
     </test>
 </suite>
